requirements.txt
ultralytics
tensorflow-cpu
keras-facenet
opencv-python-headless
numpy
fastapi
uvicorn
bcrypt
PyJWT
python-multipart
httpx
gdown
